Silence duplicated errors in queue flushed trigger
Because the `flushed` event is implemented by awaiting the last promise
in queue, it can reemit an error that has already been handled in the
regular queue cycle. This error is redundant and should be ignored.

Ideally we should rearchitecture the queue to not await the same task
twice, but for a quick bug fix this will suffice.